# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [catalog]
      summary: List categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [catalog]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [catalog]
      summary: Update category (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [catalog]
      summary: Delete category (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /brands:
    get:
      operationId: listBrands
      tags: [catalog]
      summary: List brands
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      operationId: createBrand
      tags: [catalog]
      summary: Create brand (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /brands/{id}:
    delete:
      operationId: deleteBrand
      tags: [catalog]
      summary: Delete brand (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart (auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity (auth required)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove cart item (auth required)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /addresses:
    get:
      operationId: listAddresses
      tags: [addresses]
      summary: List saved addresses (auth required)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: createAddress
      tags: [addresses]
      summary: Create address (auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /addresses/{id}:
    patch:
      operationId: updateAddress
      tags: [addresses]
      summary: Update address (auth required)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      operationId: deleteAddress
      tags: [addresses]
      summary: Delete address (auth required)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /coupons/validate:
    get:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "404":
          description: Not found
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        parentId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        description: { type: string }
        imageUrl: { type: string }
        parentId: { type: integer }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        description: { type: string }
        imageUrl: { type: string }
        parentId: { type: integer }
    Brand:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        logoUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    BrandInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        logoUrl: { type: string }
...
    CartItem:
      type: object
      required:
        [
          id,
          productId,
          productName,
          productSlug,
          productImage,
          unitPrice,
          quantity,
          color,
          size,
          stock,
        ]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        productSlug: { type: string }
        productImage: { type: ["string", "null"] }
        unitPrice: { type: number }
        quantity: { type: integer }
        color: { type: ["string", "null"] }
        size: { type: ["string", "null"] }
        stock: { type: integer }
        savedForLater: { type: boolean }
    Cart:
      type: object
      required: [items, subtotal, itemCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal: { type: number }
        itemCount: { type: integer }
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer, minimum: 1 }
        color: { type: string }
        size: { type: string }
    CartItemUpdate:
      type: object
      properties:
        quantity: { type: integer, minimum: 0 }
        savedForLater: { type: boolean }
...
    Address:
      type: object
      required:
        [
          id,
          fullName,
          phone,
          line1,
          city,
          state,
          postalCode,
          country,
          isDefault,
        ]
      properties:
        id: { type: integer }
        fullName: { type: string }
        phone: { type: string }
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }
    AddressInput:
      type: object
      required: [fullName, phone, line1, city, state, postalCode, country]
      properties:
        fullName: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        line1: { type: string, minLength: 1 }
        line2: { type: string }
        city: { type: string, minLength: 1 }
        state: { type: string, minLength: 1 }
        postalCode: { type: string, minLength: 1 }
        country: { type: string, minLength: 1 }
        isDefault: { type: boolean }
    AddressUpdate:
      type: object
      properties:
        fullName: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        line1: { type: string, minLength: 1 }
        line2: { type: string }
        city: { type: string, minLength: 1 }
        state: { type: string, minLength: 1 }
        postalCode: { type: string, minLength: 1 }
        country: { type: string, minLength: 1 }
        isDefault: { type: boolean }
    Coupon:
      type: object
      required:
        [id, code, discountType, discountValue, active, createdAt]
      properties:
        id: { type: integer }
        code: { type: string }
        discountType: { type: string, enum: [percentage, fixed] }
        discountValue: { type: number }
        minOrderAmount: { type: ["number", "null"] }
        expiresAt: { type: ["string", "null"], format: date-time }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code: { type: string, minLength: 1 }
        discountType: { type: string, enum: [percentage, fixed] }
        discountValue: { type: number, minimum: 0 }
        minOrderAmount: { type: number }
        expiresAt: { type: string, format: date-time }
        active: { type: boolean }
...