# lib/api-zod/src/generated/types/addressUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddressUpdate {
  /** @minLength 1 */
  fullName?: string;
  /** @minLength 1 */
  phone?: string;
  /** @minLength 1 */
  line1?: string;
  line2?: string;
  /** @minLength 1 */
  city?: string;
  /** @minLength 1 */
  state?: string;
  /** @minLength 1 */
  postalCode?: string;
  /** @minLength 1 */
  country?: string;
  isDefault?: boolean;
}