# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  productSlug: string;
  /** @nullable */
  productImage: string | null;
  unitPrice: number;
  quantity: number;
  /** @nullable */
  color: string | null;
  /** @nullable */
  size: string | null;
  stock: number;
  savedForLater?: boolean;
}